fix: fix remark to PR #3
