+++ src/components/Button/Button.module.scss
@@ -17,8 +17,7 @@
   padding: 0.375rem 0.75rem;
   border-radius: 0.25rem;
   text-shadow: 0px 4px 4px $font-color;
-  transition: color 0.15s ease-in-out, background-color 0.15s ease-in-out,
-    border-color 0.15s ease-in-out, box-shadow 0.15s ease-in-out;
+  transition: all 0.15s ease-in-out;
 }
 
 .btnPrimary {
