+++ src/pages/GamePage.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { PanelVoteCards } from "@/components/VoteCard/PanelVoteCards";
 import { PanelUsers } from "@/components/User/PanelUsers";
 import styles from '@/styles/GamePage.module.scss';
-import { Button } from "@/components/Button";
+import { PrimaryButton, SecondaryButton } from "@/components/Button";
 
 const GamePage = () => {
   const title = 'Name room';
@@ -24,9 +24,18 @@ const GamePage = () => {
       <div className={styles.wrapper}>
         <h1>{title}</h1>
         <div className={styles.groupBtn}>
-          <Button name="run round" myStyles="btn-primary" callback={handlerRun} />
-          <Button name="restart round" myStyles="btn-primary" callback={handlerRestart} />
-          <Button name="stop game" myStyles="btn-secondary" callback={handlerStop} />
+          <PrimaryButton
+            name="run round"
+            onClick={handlerRun}
+          />
+          <PrimaryButton
+            name="restart round"
+            onClick={handlerRestart}
+          />
+          <SecondaryButton
+            name="stop game"
+            onClick={handlerStop}
+          />
         </div>
         <PanelVoteCards />
       </div>
